Move requestCollateralTokenBalance to store/actions/blockchain PM-672
