+++ src/routes/MarketDetails/store/actions/buyShares.js
@@ -3,7 +3,7 @@ import uuid from 'uuid/v4'
 import * as api from 'api'
 import { fetchMarket, buyShares } from 'routes/MarketDetails/api'
 import Decimal from 'decimal.js'
-import { requestCollateralTokenBalance } from 'containers/HeaderContainer/store/actions'
+import { requestCollateralTokenBalance } from 'store/actions/blockchain'
 import { startLog, closeLog, closeEntrySuccess, closeEntryError } from 'routes/Transactions/store/actions/transactions'
 import { openModal, closeModal } from 'store/actions/modal'
 import { gaSend } from 'utils/analytics/google'
