+++ src/routes/MarketDetails/store/actions/sellShares.js
@@ -2,7 +2,7 @@ import allowedRangePrice from 'utils/marginPrice'
 import uuid from 'uuid/v4'
 import * as api from 'api'
 import Decimal from 'decimal.js'
-import { requestCollateralTokenBalance } from 'containers/HeaderContainer/store/actions'
+import { requestCollateralTokenBalance } from 'store/actions/blockchain'
 import { startLog, closeLog, closeEntrySuccess, closeEntryError } from 'routes/Transactions/store/actions/transactions'
 import { openModal, closeModal } from 'store/actions/modal'
 import { gaSend } from 'utils/analytics/google'
